Rayon

+++ src/forward.rs
@@ -1,6 +1,6 @@
 use crate::data_types::*;
-use crate::logsumexp::LogSumExp;
 use crate::tree::*;
+use logsumexp::LogSumExp;
 
 impl FelsensteinError {
     pub const LEAF: Self = Self::LogicError("forward_node called on a leaf");
@@ -67,7 +67,7 @@ pub fn forward_data_precompute<const DIM: usize>(
 ) where
     na::Const<DIM>: Decrementable,
     na::DefaultAllocator:
-        na::allocator::Allocator<f64, <na::Const<DIM> as na::DimSub<na::Const<1>>>::Output>,
+        na::allocator::Allocator<Float, <na::Const<DIM> as na::DimSub<na::Const<1>>>::Output>,
 {
     forward_data.log_transition.clear();
     /* What does try_symmetric_eigen() do? */
